Use same origin for API and Angular in development
In production, the browser communicates with the Angular app and API on
the same domain and port (origin), an nginx instance then serves the
static HTML and proxies the API's websocket traffic to a NodeJS
instance. The advantage of this includes: single DNS lookup, avoid
CORS (less requests and easier), easier to share cookies, single TLS
handshake, reduced TCP/UDP connections, and a single certificate.

In development, the Angular app runs on port 4250 and the API on port
5050. This means that, when the Angular app is made to communicate with
the API directly, the traffic will be cross-origin. This difference
would require different approaches to certain problems in comparison to
the production environment.

In order to keep both stacks as similar as possible, we have configured
Angular to use a proxy that automatically forwards requests that start
with /api to the api at port 5050. This essentially does the same thing
as the production environment.